refactor: split review trigger and review button
